Resolution using just getDay()
This resolution uses the getDay() function to know if it is weekend or weekdays. No need to mathematical care to deal with the difference between weekend and weekdays. Moreover, don't need a switch case to know if the client will stay more than one week. It's a way more simple resolution.