Using recursion
This method uses resursivity to count weekdays and weekends and then it calculates the costs. Furthermore, the functions that counts weekends and weekdays have a control that is the first to being called, because the client can choose any day in the week to start, not necessarily monday or saturday.